(cleanup) Move everything necessary to TopNavBar folder in components
