+++ client/components/TopNavBar/FriendItem.js
