+++ client/components/TopNavBar/FriendItemList.js
